move signing function into L2
