+++ packages/shared/src/lib/core/layer-2/utils/index.ts
@@ -14,4 +14,5 @@ export * from './getSmartContractHexName'
 export * from './parseLayer2Metadata'
 export * from './parseLayer2MetadataForTransfer'
 export * from './prepareEvmTransaction'
+export * from './signEvmTransactionWithStronghold'
 export * from './specialNativeTokenAmountEncoding'
