+++ packages/shared/src/lib/core/layer-2/utils/signEvmTransactionWithStronghold.ts
