+++ packages/shared/src/lib/core/wallet/actions/send/index.ts
@@ -4,4 +4,3 @@ export * from './sendOutputFromStardust'
 export * from './sendTransactionFromEvm'
 export * from './signAndSendEvmTransaction'
 export * from './signAndSendStardustTransaction'
-export * from './signEvmTransactionWithStronghold'
